requirements.txt
# Core utilities
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.66.0

# HuggingFace dataset streaming
datasets>=2.18.0

# Metrics
jiwer>=3.0.0

# Audio I/O + DSP
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0

# iFlytek WebSocket client
websocket-client>=1.6.0

# Google Cloud
google-cloud-speech>=2.26.0
google-cloud-texttospeech>=2.16.0
google-api-core>=2.15.0

# Gemini SDK (new)
google-genai>=0.6.0

# Azure OpenAI Python SDK
openai>=1.30.0

# Local ASR models
transformers>=4.41.0
torch>=2.1.0
sentencepiece>=0.1.99
